requirements.txt
aiohttp==3.6.2
alembic==1.4.1
altgraph==0.17
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
ColorHelper==0.1
cycler==0.10.0
dataset==1.1.2
dblpy==0.3.4
dhooks==1.1.3
discord.py==1.3.2
future==0.18.2
html5lib==1.0.1
humanize==2.0.0
idna==2.9
imgurpython==1.1.7
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.2.0
mccabe==0.6.1
mechanize==0.4.5
MouseInfo==0.1.2
multidict==4.7.5
numpy==1.18.1
pandas==1.0.2
pefile==2019.4.18
Pillow==7.0.0
praw==6.5.1
prawcore==1.0.1
PyAutoGUI==0.9.48
pycparser==2.20
PyGetWindow==0.0.8
PyInstaller==3.6
pylint==2.4.4
pymongo==3.10.1
PyMsgBox==1.0.7
PyNaCl==1.3.0
pyparsing==2.4.6
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python-editor==1.0.4
PyTweening==1.0.3
pytz==2019.3
pywin32-ctypes==0.2.0
ratelimiter==1.2.0.post0
requests==2.23.0
six==1.14.0
soupsieve==2.0
sql==0.4.0
SQLAlchemy==1.3.15
update-checker==0.16
urllib3==1.25.8
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
youtube-dl==2020.3.8
